% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wage.R
\name{get_wage}
\alias{get_wage}
\title{A description of get_wage function.}
\source{
Wages data is reproduced and distributed on
the Canada open data site and published by Employment and Social Development Canada
(Wages 2012; 2013; 2014; 2015; 2016; 2019; 2020; 2021; 2022; 2023).
}
\usage{
get_wage(dataset, api_key = Sys.getenv("CKAN_API_KEY"))
}
\arguments{
\item{dataset}{A character string or vector of character strings representing the dataset identifier(s) for wage data.}

\item{api_key}{An API key for the CKAN API. Defaults to the value stored in the \code{CKAN_API_KEY} environment variable.}
}
\value{
A tibble containing the wage data for the specified year(s).
}
\description{
This is my function to get data.
Access to Canadian wage data in Canada open data through the CKAN API
This function retrieves wage data from the Canada open data CKAN API. An API key is required for access,
and users can provide a single year or a vector of years to retrieve data for multiple years.
The function internally calls \code{get_wage_single_year} to fetch data for individual years.
}
