% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_range.R
\name{salary_range}
\alias{salary_range}
\title{Calculate Salary Range for a Specific Job and Province in a Dataset}
\usage{
salary_range(dataset, job, prov)
}
\arguments{
\item{dataset}{A tibble or data frame containing the wage dataset.}

\item{job}{A character string representing the job title for which to calculate the salary range.}

\item{prov}{A character string representing the province for which to calculate the salary range.}
}
\value{
A tibble summarizing the salary range with columns "minimum_salary" and "maximum_salary."
If no data is found for the specified job and province, NULL is returned.
}
\description{
This function calculates the salary range for a specific job and province in a given dataset.
It filters the dataset based on the provided job title and province,
then computes the average of the low and high wages to determine the salary range.
}
